api/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.9
nox==2024.4.15
pytest